+++ App_Code/Loan.cs
@@ -19,6 +19,7 @@ public class Loan : System.Web.Services.WebService {
     DataBase db = new DataBase();
     Global g = new Global();
     Settings s = new Settings();
+    User U = new User(); 
     public Loan() {
     }
 
@@ -82,16 +83,16 @@ public class Loan : System.Web.Services.WebService {
         x.includeManipulativeCosts = true;
         x.repaidOldDebt = true;
         x.buisinessUnit = new BuisinessUnit.NewUnit();
-        return JsonConvert.SerializeObject(x, Formatting.Indented);
+        return JsonConvert.SerializeObject(x, Formatting.None);
     }
 
     [WebMethod]
     public string Load(int? month, int year, string buisinessUnitCode, string search) {
         try {
             Loans xx = LoadData(month, year, buisinessUnitCode, search);
-            return JsonConvert.SerializeObject(xx, Formatting.Indented);
+            return JsonConvert.SerializeObject(xx, Formatting.None);
         } catch (Exception e) {
-            return JsonConvert.SerializeObject(e.Message, Formatting.Indented);
+            return JsonConvert.SerializeObject(e.Message, Formatting.None);
         }
     }
 
@@ -108,31 +109,30 @@ public class Loan : System.Web.Services.WebService {
                 isNewLoan = true;
             }
 
-            if (isNewLoan && x.user.restToRepayment > 0 && x.repaidOldDebt) {  // x.repaidOldDebt je za slucaj ako se greckom uplati na neciji žiro racun iako nije trazio pozajmicu, u tom slucaju treba biti false
+            if (isNewLoan && x.restToRepayment > 0 && x.repaidOldDebt) {  // x.repaidOldDebt je za slucaj ako se greskom uplati na neciji žiro racun iako nije trazio pozajmicu, u tom slucaju treba biti false
                 UpdateActiveLoan(x);  //********** Ako postoji pozajmica koja nije otplacena onda se ona otplacuje sa dijelom nove pozajmice.
             }
 
-            //TODO: Update Account manipulativeCosts & withdraw
             string sql = string.Format(@"BEGIN TRAN
-                                            IF EXISTS (SELECT * from Loan WITH (updlock,serializable) WHERE id = '{0}')
-                                                BEGIN
-                                                   UPDATE Loan SET userId = '{1}', loan = '{2}', loanDate = '{3}', repayment = '{4}', manipulativeCosts = '{5}', withdraw = '{6}', dedline = '{7}', isRepaid = '{8}', note = '{9}' WHERE id = '{0}';
-                                                   UPDATE Account SET userId = '{1}', amount = '{5}', recordDate = '{3}', mo = '{10}', yr = '{11}', note = 'Manipulativni troškovi' WHERE loanId = '{0}' AND recordType = 'manipulativeCosts';
-                                                   UPDATE Account SET userId = '{1}', amount = '{6}', recordDate = '{3}', mo = '{10}', yr = '{11}', note = 'Isplata pozajmice' WHERE loanId = '{0}' AND recordType = 'withdraw';
-                                                   UPDATE Account SET userId = '{1}', recordDate = '{3}', mo = '{10}', yr = '{11}' WHERE loanId = '{0}' AND recordType = 'loan';
-                                                END
-                                            ELSE
-                                                BEGIN
-                                                   INSERT INTO Loan (id, userId, loan, loanDate, repayment, manipulativeCosts, withdraw, dedline, isRepaid, note)
-                                                   VALUES ('{0}', '{1}', '{2}', '{3}', '{4}', '{5}', '{6}', '{7}', '{8}', '{9}')
+                                        IF EXISTS (SELECT * from Loan WITH (updlock,serializable) WHERE id = '{0}')
+                                            BEGIN
+                                                UPDATE Loan SET userId = '{1}', loan = '{2}', loanDate = '{3}', repayment = '{4}', manipulativeCosts = '{5}', withdraw = '{6}', dedline = '{7}', isRepaid = '{8}', note = '{9}' WHERE id = '{0}';
+                                                UPDATE Account SET userId = '{1}', amount = '{5}', recordDate = '{3}', mo = '{10}', yr = '{11}', note = 'Manipulativni troškovi' WHERE loanId = '{0}' AND recordType = 'manipulativeCosts';
+                                                UPDATE Account SET userId = '{1}', amount = '{6}', recordDate = '{3}', mo = '{10}', yr = '{11}', note = 'Isplata pozajmice' WHERE loanId = '{0}' AND recordType = 'withdraw';
+                                                UPDATE Account SET userId = '{1}', recordDate = '{3}', mo = '{10}', yr = '{11}' WHERE loanId = '{0}' AND recordType = 'loan';
+                                            END
+                                        ELSE
+                                            BEGIN
+                                                INSERT INTO Loan (id, userId, loan, loanDate, repayment, manipulativeCosts, withdraw, dedline, isRepaid, note)
+                                                VALUES ('{0}', '{1}', '{2}', '{3}', '{4}', '{5}', '{6}', '{7}', '{8}', '{9}')
 
-                                                   INSERT INTO Account (id, userId, amount, recordDate, mo, yr, recordType, loanId, note)
-                                                   VALUES ('{12}', '{1}', '{5}', '{3}', '{10}', '{11}', 'manipulativeCosts', '{0}', 'Manipulativni troškovi')
+                                                INSERT INTO Account (id, userId, amount, recordDate, mo, yr, recordType, loanId, note)
+                                                VALUES ('{12}', '{1}', '{5}', '{3}', '{10}', '{11}', 'manipulativeCosts', '{0}', 'Manipulativni troškovi')
 
-                                                   INSERT INTO Account (id, userId, amount, recordDate, mo, yr, recordType, loanId, note)
-                                                   VALUES ('{13}', '{1}', '{6}', '{3}', '{10}', '{11}',  'withdraw', '{0}', 'Isplata pozajmice')
-                                                END
-                                        COMMIT TRAN", x.id, x.user.id, x.loan, x.loanDate, x.repayment, x.manipulativeCosts, x.withdraw, x.dedline, x.isRepaid, x.note, g.GetMonth(x.loanDate), g.GetYear(x.loanDate), manipulativeCostsId, withdrawId);
+                                                INSERT INTO Account (id, userId, amount, recordDate, mo, yr, recordType, loanId, note)
+                                                VALUES ('{13}', '{1}', '{6}', '{3}', '{10}', '{11}',  'withdraw', '{0}', 'Isplata pozajmice')
+                                            END
+                                    COMMIT TRAN", x.id, x.user.id, x.loan, x.loanDate, x.repayment, x.manipulativeCosts, x.withdraw, x.dedline, x.isRepaid, x.note, g.GetMonth(x.loanDate), g.GetYear(x.loanDate), manipulativeCostsId, withdrawId);
             using (SqlConnection connection = new SqlConnection(g.connectionString)) {
                 connection.Open();
                 using (SqlCommand command = new SqlCommand(sql, connection)) {
@@ -140,10 +140,9 @@ public class Loan : System.Web.Services.WebService {
                 }
                 connection.Close();
             }
-
-            return JsonConvert.SerializeObject(x, Formatting.Indented);
+            return JsonConvert.SerializeObject(x, Formatting.None);
         } catch (Exception e) {
-            return JsonConvert.SerializeObject("Error: " + e.Message, Formatting.Indented);
+            return JsonConvert.SerializeObject("Error: " + e.Message, Formatting.None);
         }
     }
 
@@ -168,55 +167,61 @@ public class Loan : System.Web.Services.WebService {
                 }
                 connection.Close();
             }
-            return JsonConvert.SerializeObject(x, Formatting.Indented);
+            return JsonConvert.SerializeObject(x, Formatting.None);
         } catch (Exception e) {
-            return JsonConvert.SerializeObject(e.Message, Formatting.Indented);
+            return JsonConvert.SerializeObject(e.Message, Formatting.None);
         }
     }
 
     [WebMethod]
     public string Delete(string id) {
         try {
-            string sql = string.Format(@"DELETE FROM Loan WHERE id = '{0}';
+            string msg = null;
+            if (!string.IsNullOrEmpty(id)) {
+                string sql = string.Format(@"DELETE FROM Loan WHERE id = '{0}';
                                         DELETE FROM Account WHERE loanId = '{0}'", id);
-            using (SqlConnection connection = new SqlConnection(g.connectionString)) {
-                connection.Open();
-                using (SqlCommand command = new SqlCommand(sql, connection)) {
-                    command.ExecuteReader();
+                using (SqlConnection connection = new SqlConnection(g.connectionString)) {
+                    connection.Open();
+                    using (SqlCommand command = new SqlCommand(sql, connection)) {
+                        command.ExecuteReader();
+                    }
+                    connection.Close();
                 }
-                connection.Close();
+                msg = "Obrisano";
+            } else {
+                msg = "Pozajmica ne može biti obrisana jer nije spremljena";
             }
-            return JsonConvert.SerializeObject("Obrisano", Formatting.Indented);
+            return JsonConvert.SerializeObject(msg, Formatting.None);
         } catch (Exception e) {
-            return JsonConvert.SerializeObject(e.Message, Formatting.Indented);
+            return JsonConvert.SerializeObject(e.Message, Formatting.None);
         }
     }
 
     public Loans LoadData(int? month, int year, string buisinessUnitCode, string search) {
         db.Loan();
-            string sql = string.Format(@"SELECT l.id, l.userId, l.loan, l.loanDate, l.repayment, l.manipulativeCosts, l.withdraw, l.dedline, l.isRepaid, l.note, u.firstName, u.lastName, b.code, b.title FROM Loan l
-                        LEFT OUTER JOIN Users u
-                        ON l.userId = u.id
-                        LEFT OUTER JOIN BuisinessUnit b
-                        ON u.buisinessUnitCode = b.code
-                        WHERE {0} {1} {2}"
-                            , string.Format("CONVERT(datetime, l.loanDate) >= '{0}' AND CONVERT(datetime, l.loanDate) < '{1}'", g.ReffDate(month == null ? 1 : month, year), g.ReffDate(month == null ? 12 : month + 1, year))
-                            , string.IsNullOrEmpty(buisinessUnitCode) ? "" : string.Format("AND u.buisinessUnitCode = '{0}'", buisinessUnitCode)
-                            , string.IsNullOrEmpty(search) ? "" : string.Format("AND u.id LIKE '%{0}%' OR u.firstName LIKE N'{0}%' OR u.lastName LIKE N'{0}%'", search));
-            Loans xx = new Loans();
-            xx.data = new List<NewLoan>();
-            using (SqlConnection connection = new SqlConnection(g.connectionString)) {
-                connection.Open();
-                using (SqlCommand command = new SqlCommand(sql, connection)) {
-                    using (SqlDataReader reader = command.ExecuteReader()) {
-                        while (reader.Read()) {
-                            NewLoan x = ReadData(reader);
-                            xx.data.Add(x);
-                        }
+        string sql = string.Format(@"SELECT l.id, l.userId, l.loan, l.loanDate, l.repayment, l.manipulativeCosts, l.withdraw, l.dedline, l.isRepaid, l.note, u.firstName, u.lastName, b.code, b.title FROM Loan l
+                    LEFT OUTER JOIN Users u
+                    ON l.userId = u.id
+                    LEFT OUTER JOIN BuisinessUnit b
+                    ON u.buisinessUnitCode = b.code
+                    WHERE {0} {1} {2}"
+                        , string.Format("CONVERT(datetime, l.loanDate) >= '{0}' AND CONVERT(datetime, l.loanDate) < '{1}'", g.ReffDate(month == null ? 1 : month, year), g.ReffDate(month == null ? 12 : month + 1, year))
+                        , string.IsNullOrEmpty(buisinessUnitCode) ? "" : string.Format("AND u.buisinessUnitCode = '{0}'", buisinessUnitCode)
+                        , string.IsNullOrEmpty(search) ? "" : string.Format("AND u.id LIKE '%{0}%' OR u.firstName LIKE N'{0}%' OR u.lastName LIKE N'{0}%'", search));
+        Loans xx = new Loans();
+        xx.data = new List<NewLoan>();
+        using (SqlConnection connection = new SqlConnection(g.connectionString)) {
+            connection.Open();
+            using (SqlCommand command = new SqlCommand(sql, connection)) {
+                using (SqlDataReader reader = command.ExecuteReader()) {
+                    while (reader.Read()) {
+                        NewLoan x = ReadData(reader);
+                        xx.data.Add(x);
                     }
                 }
-                connection.Close();
             }
+            connection.Close();
+        }
 
         xx.data = xx.data.OrderBy(a => a.user.lastName).ToList();
         xx.total = new Total();
@@ -237,6 +242,8 @@ public class Loan : System.Web.Services.WebService {
         x.id = reader.GetValue(0) == DBNull.Value ? null : reader.GetString(0);
         x.user = new User.NewUser();
         x.user.id = reader.GetValue(1) == DBNull.Value ? null : reader.GetString(1);
+
+        x.user = U.GetUserData(x.user.id, null);
         x.loan = reader.GetValue(2) == DBNull.Value ? 0 : Convert.ToDouble(reader.GetString(2));
         x.loanDate = reader.GetValue(3) == DBNull.Value ? null : reader.GetString(3);
         x.repayment = reader.GetValue(4) == DBNull.Value ? 0 : Convert.ToDouble(reader.GetString(4));
@@ -257,6 +264,32 @@ public class Loan : System.Web.Services.WebService {
         x.user.buisinessUnit.title = reader.GetValue(13) == DBNull.Value ? null : reader.GetString(13);
         return x;
     }
+
+    //NewLoan ReadData(SqlDataReader reader) {
+    //    NewLoan x = new NewLoan();
+    //    x.id = reader.GetValue(0) == DBNull.Value ? null : reader.GetString(0);
+    //    x.user = new User.NewUser();
+    //    x.user.id = reader.GetValue(1) == DBNull.Value ? null : reader.GetString(1);
+    //    x.loan = reader.GetValue(2) == DBNull.Value ? 0 : Convert.ToDouble(reader.GetString(2));
+    //    x.loanDate = reader.GetValue(3) == DBNull.Value ? null : reader.GetString(3);
+    //    x.repayment = reader.GetValue(4) == DBNull.Value ? 0 : Convert.ToDouble(reader.GetString(4));
+    //    x.manipulativeCosts = reader.GetValue(5) == DBNull.Value ? 0 : Convert.ToDouble(reader.GetString(5));
+    //    x.manipulativeCostsCoeff = s.Data().manipulativeCostsCoeff;
+    //    x.actualLoan = x.loan - x.manipulativeCosts;
+    //    x.withdraw = reader.GetValue(6) == DBNull.Value ? 0 : Convert.ToDouble(reader.GetString(6));
+    //    x.lastLoanToRepaid = x.actualLoan - x.withdraw;
+    //    User u = new User();
+    //    x.restToRepayment = GetTotalLoan(x) - GetTotalLoanRepayed(x);  // TOOD: provjeriti
+    //    x.dedline = reader.GetValue(7) == DBNull.Value ? s.Data().defaultDedline : Convert.ToDouble(reader.GetString(7));
+    //    x.isRepaid = reader.GetValue(8) == DBNull.Value ? 0 : reader.GetInt32(8);
+    //    x.note = reader.GetValue(9) == DBNull.Value ? null : reader.GetString(9);
+    //    x.user.firstName = reader.GetValue(10) == DBNull.Value ? null : reader.GetString(10);
+    //    x.user.lastName = reader.GetValue(11) == DBNull.Value ? null : reader.GetString(11);
+    //    x.user.buisinessUnit = new BuisinessUnit.NewUnit();
+    //    x.user.buisinessUnit.code = reader.GetValue(12) == DBNull.Value ? null : reader.GetString(12);
+    //    x.user.buisinessUnit.title = reader.GetValue(13) == DBNull.Value ? null : reader.GetString(13);
+    //    return x;
+    //}
 
     private List<MonthlyTotal> GetMonthlyTotal(List<NewLoan> data) {
         List<MonthlyTotal> xx = new List<MonthlyTotal>();
@@ -313,14 +346,13 @@ public class Loan : System.Web.Services.WebService {
             connection.Close();
         }
 
-        //U Account tbl uplati razliku za otplatu
         //sql = string.Format(@"BEGIN TRAN
         //                        BEGIN
         //                            INSERT INTO Account (id, userId, amount, recordDate, mo, yr, recordType, loanId, note)
         //                            VALUES ('{0}', '{1}', '{2}', '{3}', '{4}', '{5}', '{6}', '{7}', '{8}')
-        //                            UPDATE Loan SET isRepaid = 1 WHERE id = '{7}'
+        //                            UPDATE Loan SET isRepaid = 1 WHERE id = '{9}'
         //                        END
-        //                    COMMIT TRAN", Guid.NewGuid().ToString(), x.user.id, x.user.restToRepayment, x.loanDate, g.GetMonth(x.loanDate), g.GetYear(x.loanDate), g.repayment, loanId, "Otplata novom pozajmicom");
+        //                    COMMIT TRAN", Guid.NewGuid().ToString(), x.user.id, x.user.restToRepayment, x.loanDate, g.GetMonth(x.loanDate), g.GetYear(x.loanDate), g.loan, x.id, "Otplata novom pozajmicom", loanId);
 
         sql = string.Format(@"BEGIN TRAN
                                 BEGIN
@@ -328,7 +360,7 @@ public class Loan : System.Web.Services.WebService {
                                     VALUES ('{0}', '{1}', '{2}', '{3}', '{4}', '{5}', '{6}', '{7}', '{8}')
                                     UPDATE Loan SET isRepaid = 1 WHERE id = '{9}'
                                 END
-                            COMMIT TRAN", Guid.NewGuid().ToString(), x.user.id, x.user.restToRepayment, x.loanDate, g.GetMonth(x.loanDate), g.GetYear(x.loanDate), g.loan, x.id, "Otplata novom pozajmicom", loanId);
+                            COMMIT TRAN", Guid.NewGuid().ToString(), x.user.id, x.restToRepayment, x.loanDate, g.GetMonth(x.loanDate), g.GetYear(x.loanDate), g.loan, x.id, "Otplata novom pozajmicom", loanId);
 
         using (SqlConnection connection = new SqlConnection(g.connectionString)) {
             connection.Open();
@@ -337,21 +369,6 @@ public class Loan : System.Web.Services.WebService {
             }
             connection.Close();
         }
-
-        //U Loan tbl stavi isRepaid = 1
-        //sql = string.Format(@"BEGIN TRAN
-        //                        BEGIN
-        //                            UPDATE Loan SET isRepaid = 1 WHERE id = '{0}'
-        //                        END
-        //                    COMMIT TRAN", x.id);
-        //using (SqlConnection connection = new SqlConnection(connectionString)) {
-        //    connection.Open();
-        //    using (SqlCommand command = new SqlCommand(sql, connection)) {
-        //        command.ExecuteNonQuery();
-        //    }
-        //    connection.Close();
-        //}
-
 
     }
